pipeline/requirements.txt
anthropic>=0.40.0
watchdog>=4.0.0
Pillow>=10.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
